+++ src/dexterity/localroles/tests/test_utility.py
@@ -3,7 +3,7 @@ import unittest2 as unittest
 from zope.component import getUtility
 from plone.app.testing import login, TEST_USER_NAME, setRoles, TEST_USER_ID
 
-from ..utility import ParentRelatedSearch
+from ..utility import ParentRelatedSearch, runRelatedSearch
 from ..interfaces import ILocalRolesRelatedSearchUtility
 from ..testing import DLR_PROFILE_FUNCTIONAL
 
@@ -24,3 +24,10 @@ class TestRelatedSearchUtility(unittest.TestCase):
         utility = getUtility(ILocalRolesRelatedSearchUtility, 'dexterity.localroles.related_parent')
         self.assertIsInstance(utility, ParentRelatedSearch)
         self.assertEqual(utility.get_objects(item), [self.portal])
+
+    def test_runRelatedSearch(self):
+        self.portal.invokeFactory('testingtype', 'test')
+        item = self.portal.get('test')
+        self.assertEqual(runRelatedSearch('error', item), [])
+        self.assertEqual(runRelatedSearch('dexterity.localroles.related_parent', item),
+                         [self.portal])