+++ src/dexterity/localroles/tests/test_utils.py
@@ -3,7 +3,9 @@ import unittest2 as unittest
 from plone.app.testing import login, TEST_USER_NAME, setRoles, TEST_USER_ID
 
 from ..testing import DLR_PROFILE_FUNCTIONAL
-from ..utils import add_fti_configuration, fti_configuration
+from ..utils import (add_related_roles, del_related_roles, get_related_roles, set_related_roles,
+                     fti_configuration, add_fti_configuration)
+
 localroles_config = {
     u'private': {'raptor': {'roles': ('Editor', 'Contributor')}, 'cavemans': {'roles': ('Reader', )}},
     u'published': {'hunters': {'roles': ('Reader',)}, 'dina': {'roles': ('Editor',)}}}
@@ -18,6 +20,39 @@ class TestUtils(unittest.TestCase):
         self.portal = self.layer['portal']
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
         login(self.portal, TEST_USER_NAME)
+
+    def test_add_related_roles(self):
+        # nothing already stored
+        add_related_roles(self.portal, 'fakeuid', 'raptor', ['Reader'])
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'), {'raptor': set(['Reader'])})
+        # adding to existing
+        add_related_roles(self.portal, 'fakeuid', 'raptor', ['Reviewer'])
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'), {'raptor': set(['Reader', 'Reviewer'])})
+        # adding another
+        add_related_roles(self.portal, 'fakeuid', 't-rex', ['Reader'])
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'),
+                             {'raptor': set(['Reader', 'Reviewer']), 't-rex': set(['Reader'])})
+
+    def test_del_related_roles(self):
+        # no key or no uid stored
+        self.assertDictEqual(del_related_roles(self.portal, 'fakeuid'), {})
+        # adding value
+        add_related_roles(self.portal, 'fakeuid', 'raptor', ['Reader'])
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'), {'raptor': set(['Reader'])})
+        self.assertDictEqual(del_related_roles(self.portal, 'fakeuid'), {'raptor': set(['Reader'])})
+        self.assertDictEqual(del_related_roles(self.portal, 'fakeuid'), {})
+
+    def test_get_related_roles(self):
+        # no key or no uid stored
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'), {})
+
+    def test_set_related_roles(self):
+        # nothing already stored
+        set_related_roles(self.portal, 'fakeuid', {'raptor': set(['Reader'])})
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'), {'raptor': set(['Reader'])})
+        # replacing existing
+        set_related_roles(self.portal, 'fakeuid', {'t-rex': set(['Reader'])})
+        self.assertDictEqual(get_related_roles(self.portal, 'fakeuid'), {'t-rex': set(['Reader'])})
 
     def test_fti_configuration(self):
         self.portal.invokeFactory('Document', 'doc')