+++ src/dexterity/localroles/utility.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 """Example."""
 from Acquisition import aq_inner, aq_parent
+from zope.component import getUtility
+from zope.component.interfaces import ComponentLookupError
 from zope.interface import implements
 
 from .interfaces import ILocalRolesRelatedSearchUtility
@@ -14,3 +16,12 @@ class ParentRelatedSearch(object):
     def get_objects(self, context):
         """ Return the parent. """
         return [aq_parent(aq_inner(context))]
+
+
+def runRelatedSearch(utility, context):
+    """ Run the related search and return the result """
+    try:
+        util = getUtility(ILocalRolesRelatedSearchUtility, utility)
+    except ComponentLookupError:
+        return []
+    return util.get_objects(context)